Created basic app config colors and file
